requirements.txt
Flask
flask-mqtt
Flask-InfluxDB
flask_sqlalchemy
Flask-Migrate